# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateSpotPriceRequest:
    properties:
      currency_id:
        type: string
      price:
        example: 42.5
        type: number
      timestamp:
        example: "2024-03-20T13:00:00Z"
        type: string
      zone_id:
        type: string
    required:
    - currency_id
    - price
    - timestamp
    - zone_id
    type: object
  models.CreateSpotPricesRequest:
    properties:
      spot_prices:
        items:
          $ref: '#/definitions/models.CreateSpotPriceRequest'
        minItems: 1
        type: array
    required:
    - spot_prices
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.SpotPrice:
    properties:
      created_at:
        type: string
      currency_id:
        type: string
      id:
        type: string
      price:
        type: number
      timestamp:
        type: string
      updated_at:
        type: string
      zone_id:
        type: string
    required:
    - currency_id
    - price
    - timestamp
    - zone_id
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: WattWatch API server.
  title: WattWatch API
  version: "1.0"
paths:
...
  /spot-prices:
    get:
      consumes:
      - application/json
      description: Returns a list of spot prices for a specific zone and currency
        within a date range (max 7 days)
      parameters:
      - description: Zone name (e.g., 'SE1')
        in: query
        name: zone
        required: true
        type: string
      - description: Currency name (e.g., 'EUR')
        in: query
        name: currency
        required: true
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        required: true
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        required: true
        type: string
      - description: Order descending
        in: query
        name: order_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SpotPrice'
            type: array
        "400":
          description: Invalid parameters or date range exceeds 7 days
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Zone or currency not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List spot prices
      tags:
      - spot-prices
    post:
      consumes:
      - application/json
      description: Creates or updates one or more spot prices in a single transaction.
        If a spot price with the same timestamp, zone_id, and currency_id exists,
        its price will be updated. Requires admin privileges.
      parameters:
      - description: Spot prices to create or update
        in: body
        name: spot_prices
        required: true
        schema:
          $ref: '#/definitions/models.CreateSpotPricesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.SpotPrice'
            type: array
        "400":
          description: Invalid request body, negative price, or invalid zone/currency
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create or update spot prices (Admin only)
      tags:
      - spot-prices
  /spot-prices/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing spot price. Requires admin privileges.
      parameters:
      - description: Spot Price ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Spot price deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid spot price ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Spot price not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a spot price (Admin only)
      tags:
      - spot-prices
    get:
      consumes:
      - application/json
      description: Returns a spot price by its ID
      parameters:
      - description: Spot Price ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SpotPrice'
        "400":
          description: Invalid spot price ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Spot price not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a spot price by ID
      tags:
      - spot-prices
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"